chore: Move files to sendEmail module
